rename phase1 to phase2 and refactor test
